+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed issue of SLM portal login for iSCSI backends (Issue [#387](https://github.com/NetApp/trident/issues/387)).
 - Fixed issue where Trident would crash when creating a volume if ONTAP FCP interfaces were present in the target SVM.
 - Fixed issue where storage prefix with period would cause backend to fail when updating or creating backend or upgrading Trident from v20.01
+- Fixed ONTAP SAN drivers to clean up if provisioning fails (Issue [#442](https://github.com/NetApp/trident/issues/442)).
 
 **Enhancements:**
 - **Kubernetes:** Added automatic node preparation for NFS protocol. By default, Trident will now attempt to make